move pack generation to cubefn.js
